+++ app/src/main/AndroidManifest.xml
@@ -11,17 +11,16 @@
         android:supportsRtl="false"
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".MainActivity">
-        </activity>
-
-        <activity android:name=".ItemActivity">
+        <activity android:name=".MainActivity"></activity>
+        <activity android:name=".ItemsFragment">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-<!-- Приложение запускается с ItemActivity -->
+            <!-- Приложение запускается с ItemActivity -->
         </activity>
+        <activity android:name=".AddActivity"></activity>
     </application>
 
 </manifest>
