+++ app/src/main/java/com/loftschool/moneytracker/ItemsFragment.java
@@ -1,0 +1,29 @@
+package com.loftschool.moneytracker;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.support.v7.widget.LinearLayoutManager;
+import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+// Новый активити со списком трат (item-ов)
+public class ItemsFragment extends Fragment { //наследуется от Fragment, так как это фрагмент
+
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) { //При создании вьюшки берем стиль xml
+        View view = inflater.inflate(R.layout.fragment_items,container,false);
+        return view; //возвращаем вьюшку
+    }
+
+    @Override
+    public void onViewCreated(View view, @Nullable Bundle savedInstanceState) { //Когда вьюшка создана
+        RecyclerView recycler = view.findViewById(R.id.recycler); //использовать RecyclerView по айди из xml файла
+        recycler.setLayoutManager(new LinearLayoutManager(getContext())); // Задать LayoutManager
+        recycler.setAdapter(new ItemAdapter()); //Использовать созданный адаптер
+    }
+}
