+++ app/src/main/res/layout/activity_main.xml
@@ -7,58 +7,4 @@
     android:background="@drawable/bg_pattern"
     tools:context="com.loftschool.moneytracker.MainActivity">
 
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@android:color/white"
-        android:orientation="horizontal"
-        android:padding="@dimen/padding_name">
-
-
-        <EditText
-            android:id="@+id/name"
-            android:backgroundTint="@color/colorZeroAlpha"
-            android:layout_width="0dp"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:hint="@string/name_string"
-            android:inputType="text"
-            android:maxLines="1" />
-
-        <EditText
-            android:id="@+id/money"
-            android:backgroundTint="@color/colorZeroAlpha"
-            android:layout_width="@dimen/money_dim"
-            android:layout_height="wrap_content"
-            android:gravity="end"
-            android:hint="@string/money_string"
-            android:inputType="number"
-            android:maxLines="1" />
-
-        <ImageButton
-            android:id="@+id/add"
-            android:background="@color/colorInactiveButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center_vertical"
-            android:contentDescription="@string/add_to_base_string"
-            android:src="@drawable/ic_add_black_24dp" />
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="74dp">
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:alpha="0.5"
-            android:src="@drawable/shadow"
-            tools:ignore="ContentDescription" />
-    </LinearLayout>
-
-
 </FrameLayout>
