+++ app/src/main/res/layout/fragment_items.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.loftschool.moneytracker.ItemActivity">
+    tools:context="com.loftschool.moneytracker.ItemsFragment">
 
 
     <android.support.v7.widget.RecyclerView
